Example views: fixed_brunelhand, fixed_crazyflie
Photographs of workspace deployments, from perspectives that are
outside the workspace and that show all hardware except for host
computers.

A *host computer* is one to which root access is provided as part of
an instance and to which some or all of the robot hardware (sensors
etc.) of the respective workspace type are attached.